Code cleanup to eliminate warnings
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6198 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 9f9b8b46ff5eb063dc992e0e67728625d7f27567
Former-commit-id: a181913faee8558cb541a10c4c050657b20d3a52

+++ src/drivers/kilo2008/kdriver.cpp
@@ -1635,13 +1635,13 @@ vec2f KDriver::TargetPoint() {
     n.normalize();
     vec2f d((seg->vertex[TR_EL].x - seg->vertex[TR_SL].x) / seg->length,
       (seg->vertex[TR_EL].y - seg->vertex[TR_SL].y) / seg->length);
-    return s + d * length + static_cast<float>(offset) * n;
+    return (s + d * (float) length + static_cast<float>((float) offset) * n);
   } else {
     vec2f c(seg->center.x, seg->center.y);
     double arc = length / seg->radius;
     double arcsign = (seg->type == TR_RGT) ? -1.0 : 1.0;
     arc = arc * arcsign;
-    s = s.rotate(c, arc);
+    s = s.rotate(c, (float) arc);
 
     vec2f n, t, rt;
     n = c - s;
